cfr test: move refactor split
Move the controller forced reconfiguration tests into their own
tests/rptest/cfr_tests/ directory.

Extract the shared base class
(ControllerForceReconfigurationTestBase) into cfr_test_base.py.

This prepares for the upcoming CLI-script CFR tests to reuse the same
cluster lifecycle helpers (recovery mode toggling, majority killing,
node joining, etc.) without duplicating code.
