+++ tests/rptest/test_suite_quick.yml
@@ -11,6 +11,7 @@ quick:
   included:
   - tests/
   - transactions/
+  - cfr_tests/
   - direct_consumer_tests/
 
   excluded:
